+++ src/core.js
@@ -1,17 +1,11 @@
 goog.provide('olcs.core');
 
-goog.require('goog.Promise');
 goog.require('goog.asserts');
-goog.require('goog.async.AnimationDelay');
 goog.require('ol.layer.Tile');
 goog.require('ol.proj');
 goog.require('ol.source.TileImage');
 goog.require('ol.source.WMTS');
 goog.require('olcs.core.OLImageryProvider');
-
-// Initialize rejection handler.
-// Do nothing on rejected promise, like with standard Promises.
-goog.Promise.setUnhandledRejectionHandler(function() {});
 
 
 /**
@@ -86,7 +80,6 @@ olcs.core.applyHeightOffsetToGeometry = function(geometry, height) {
  * @param {!Cesium.Cartesian3} axis
  * @param {!Cesium.Matrix4} transform
  * @param {olcsx.core.RotateAroundAxisOption=} opt_options
- * @return {goog.Promise}
  * @api
  */
 olcs.core.rotateAroundAxis = function(camera, angle, axis, transform,
@@ -99,33 +92,32 @@ olcs.core.rotateAroundAxis = function(camera, angle, axis, transform,
   var easing = defaultValue(options.easing, ol.easing.linear);
   var callback = options.callback;
 
-  var start = Date.now();
   var lastProgress = 0;
   var oldTransform = new Cesium.Matrix4();
 
-  return new goog.Promise(function(resolve, reject) {
-    var animation = new goog.async.AnimationDelay(function(millis) {
-      var progress = easing(clamp((millis - start) / duration, 0, 1));
-      goog.asserts.assert(progress > lastProgress);
+  var start = Date.now();
+  var step = function() {
+    var timestamp = Date.now();
+    var timeDifference = timestamp - start;
+    var progress = easing(clamp(timeDifference / duration, 0, 1));
+    goog.asserts.assert(progress >= lastProgress);
 
-      camera.transform.clone(oldTransform);
-      var stepAngle = (progress - lastProgress) * angle;
-      lastProgress = progress;
-      camera.lookAtTransform(transform);
-      camera.rotate(axis, stepAngle);
-      camera.lookAtTransform(oldTransform);
+    camera.transform.clone(oldTransform);
+    var stepAngle = (progress - lastProgress) * angle;
+    lastProgress = progress;
+    camera.lookAtTransform(transform);
+    camera.rotate(axis, stepAngle);
+    camera.lookAtTransform(oldTransform);
 
-      if (progress < 1) {
-        animation.start();
-      } else {
-        if (callback) {
-          callback();
-        }
-        resolve();
+    if (progress < 1) {
+      window.requestAnimationFrame(step);
+    } else {
+      if (callback) {
+        callback();
       }
-    });
-    animation.start();
-  });
+    }
+  };
+  window.requestAnimationFrame(step);
 };